Support setting other defaults in the re-usable agent besides cookies.
Since `superagent` 3.8, there has been support for the re-usable agent
supporting other defaults besides cookies.

For example, it's particularly useful to set default 'Authorization'
header before running repeated tests against an authenticated API.

I tracked down which version of `superagent` we needed to use as a
minimum dependency to this commit:

https://github.com/visionmedia/superagent/commit/66aed3427f29bf0c09b1df0e7f8aa9f0e30732ab

This update should be considered a bug fix, as 'supertest' claims to
support all of superagent's features, but this feature is documented in
superagent, but wasn't working in `supertest'.